Move selectors into their own folder
