+++ static/app/views/starfish/views/spans/selectors/domainSelector.tsx
