+++ static/app/views/starfish/views/spans/selectors/spanOperationSelector.tsx
